+++ lib/wechat_pay.rb
@@ -3,6 +3,7 @@ require 'wechat_pay/access_token'
 require 'wechat_pay/app'
 require 'wechat_pay/sign'
 require 'wechat_pay/utils'
+require 'wechat_pay/notify'
 
 module WechatPay
   class << self