+++ spec/wechat_pay/notify_spec.rb
@@ -1,0 +1,9 @@
+require 'spec_helper'
+
+describe WechatPay::Notify do
+  describe ".verify?" do
+    it "with nil sign" do
+      WechatPay::Notify.verify?({}).must_equal(false)
+    end
+  end
+end